Hoist standalone types out of AppModel.swift
First step of the AppModel decomposition mandated by Docs/Architecture.md.
Move the types that never touched AppModel state into their own files:
the app lifecycle crash journal, the automatic-inspector prompt and runs
navigation values (into Models/InspectorModels.swift beside InspectorTab),
the command palette catalog, the display-synchronized flush driver, and
String.nilIfEmpty. The private JSON bridging helpers become module-scope
functions so later feature models can share them; the one self-qualified
call site drops its 'self.'.

Behavior-identical refactor; project regenerated with xcodegen.
