+++ Locus/AppModel.swift
@@ -43,108 +43,6 @@ private final class ChatExportAccessoryView: NSView {
     }
 }
 
-struct AppLifecycleRunSnapshot: Codable, Equatable {
-    let sessionID: String
-    let runID: String
-    let state: TeamRunState
-    let updatedAt: Date
-}
-
-struct AppLifecycleRecovery: Equatable {
-    let snapshot: AppLifecycleRunSnapshot?
-
-    var message: String {
-        guard let snapshot else {
-            return "Locus did not close normally. Your last session was restored."
-        }
-        switch snapshot.state {
-        case .completed:
-            return "Locus was force quit after the team run completed. Its results were restored."
-        case .failed, .cancelled, .discarded:
-            return "Locus did not close normally. The last team run finished as \(snapshot.state.title.lowercased())."
-        case .interrupted:
-            return "Locus closed unexpectedly. The interrupted team run is ready to resume."
-        case .queued, .dispatching, .running, .waitingPermission, .waitingComputer,
-             .waitingDispatchApproval, .reviewing, .paused:
-            return "Locus closed unexpectedly. The last team run can be inspected and resumed."
-        }
-    }
-}
-
-/// A deliberately tiny crash journal. The durable session/run database remains
-/// authoritative; these defaults only tell the next launch what to reopen and
-/// whether the previous process reached its ordinary termination hook.
-final class AppLifecycleJournal {
-    private let defaults: UserDefaults
-    private let cleanKey: String
-    private let snapshotKey: String
-
-    init(defaults: UserDefaults = .standard, keyPrefix: String = "Locus.lifecycle") {
-        self.defaults = defaults
-        cleanKey = "\(keyPrefix).clean"
-        snapshotKey = "\(keyPrefix).latestRun"
-    }
-
-    @discardableResult
-    func beginLaunch() -> AppLifecycleRecovery? {
-        let hadPreviousLaunch = defaults.object(forKey: cleanKey) != nil
-        let previousLaunchWasClean = defaults.bool(forKey: cleanKey)
-        let snapshot = defaults.data(forKey: snapshotKey)
-            .flatMap { try? JSONDecoder().decode(AppLifecycleRunSnapshot.self, from: $0) }
-        // Set this before any services are started. Force Quit cannot run a
-        // callback, so leaving this false is the abnormal-exit signal.
-        defaults.set(false, forKey: cleanKey)
-        guard hadPreviousLaunch, !previousLaunchWasClean else { return nil }
-        return AppLifecycleRecovery(snapshot: snapshot)
-    }
-
-    func record(sessionID: String, runID: String, state: TeamRunState, at date: Date = Date()) {
-        guard !sessionID.isEmpty, !runID.isEmpty else { return }
-        let snapshot = AppLifecycleRunSnapshot(
-            sessionID: sessionID,
-            runID: runID,
-            state: state,
-            updatedAt: date
-        )
-        if let data = try? JSONEncoder().encode(snapshot) {
-            defaults.set(data, forKey: snapshotKey)
-        }
-    }
-
-    func markCleanExit() {
-        defaults.set(true, forKey: cleanKey)
-    }
-}
-
-struct AutomaticInspectorPrompt: Equatable {
-    let tab: InspectorTab
-    let runID: String?
-
-    var isTeamRun: Bool { tab == .runs }
-
-    var title: String {
-        isTeamRun
-            ? "Open Runs for team requests?"
-            : "Open Context & Plan for solo requests?"
-    }
-
-    var message: String {
-        if isTeamRun {
-            return "Locus can open Runs whenever you send a team request so you can follow its agents and progress. You can change this anytime in Settings → General → Conversation."
-        }
-        return "Locus can open Context & Plan whenever you send a solo Work request so you can follow context use and the current plan. You can change this anytime in Settings → General → Conversation."
-    }
-
-    var confirmationTitle: String {
-        isTeamRun ? "Open Runs Every Time" : "Open Context & Plan Every Time"
-    }
-}
-
-struct RunsNavigationRequest: Equatable, Identifiable {
-    let id = UUID()
-    let runID: String
-}
-
 @MainActor
 final class AppModel: ObservableObject {
     enum ProviderConnectionTestFollowUp: Equatable {
@@ -3788,7 +3686,7 @@ final class AppModel: ObservableObject {
                 "text": payload,
                 "mode": dispatchedMode.rawValue,
             ]
-            if let agentConfig = self.encodedJSONObject(self.primaryAgentBehavior) {
+            if let agentConfig = encodedJSONObject(self.primaryAgentBehavior) {
                 request["agent_config"] = agentConfig
             }
             if let dispatchedTeam { request["team"] = dispatchedTeam }
@@ -15223,18 +15121,6 @@ final class AppModel: ObservableObject {
         }
     }
 
-    private func decode<T: Decodable>(_ type: T.Type, from object: [String: Any]) -> T? {
-        guard JSONSerialization.isValidJSONObject(object),
-              let data = try? JSONSerialization.data(withJSONObject: object)
-        else { return nil }
-        return try? JSONDecoder().decode(type, from: data)
-    }
-
-    private func encodedJSONObject<T: Encodable>(_ value: T) -> [String: Any]? {
-        guard let data = try? JSONEncoder().encode(value) else { return nil }
-        return try? JSONSerialization.jsonObject(with: data) as? [String: Any]
-    }
-
     nonisolated private static func readChatAttachments(
         _ selected: [URL],
         excluding existing: Set<URL>
@@ -16345,201 +16231,6 @@ extension AppModel {
     }
 }
 
-enum CommandAction: String, CaseIterable, Identifiable {
-    case newSession
-    case clearChat
-    case clearSessions
-    case reviewChanges
-    case createCheckpoint
-    case askMode
-    case workMode
-    case planMode
-    case grillMode
-    case chooseWorkspace
-    case newWorkspace
-    case browseModels
-    case refreshModels
-    case exportSession
-    case permissions
-    case searchConversations
-    case showUsage
-    case showShortcuts
-    case showNotebook
-    case openSettings
-
-    var id: String { rawValue }
-
-    var title: String {
-        switch self {
-        case .newSession: "Start a new session"
-        case .clearChat: "Clear chat"
-        case .clearSessions: "Clear saved sessions"
-        case .reviewChanges: "Review file changes"
-        case .createCheckpoint: "Create a session checkpoint"
-        case .askMode: "Turn on Just Chat"
-        case .workMode: "Use adaptive Work mode"
-        case .planMode: "Switch to Plan mode"
-        case .grillMode: "Switch to Grill mode"
-        case .chooseWorkspace: "Choose a workspace"
-        case .newWorkspace: "Create a new workspace folder"
-        case .browseModels: "Browse Hugging Face models"
-        case .refreshModels: "Refresh installed models"
-        case .exportSession: "Export current session as Markdown"
-        case .permissions: "Change what the agent may do without asking"
-        case .searchConversations: "Search all conversations"
-        case .showUsage: "Show usage and costs"
-        case .showShortcuts: "Show keyboard shortcuts"
-        case .showNotebook: "Open the notebook"
-        case .openSettings: "Open Settings"
-        }
-    }
-
-    var symbol: String {
-        switch self {
-        case .newSession: "plus"
-        case .clearChat: "eraser"
-        case .clearSessions: "trash"
-        case .reviewChanges: "doc.text.magnifyingglass"
-        case .createCheckpoint: "clock.arrow.circlepath"
-        case .askMode: "bubble.left"
-        case .workMode: "sparkles"
-        case .planMode: "list.bullet.clipboard"
-        case .grillMode: "flame"
-        case .chooseWorkspace: "folder"
-        case .newWorkspace: "folder.badge.plus"
-        case .browseModels: "shippingbox.and.arrow.backward"
-        case .refreshModels: "arrow.clockwise"
-        case .exportSession: "square.and.arrow.up"
-        case .permissions: "shield.lefthalf.filled"
-        case .searchConversations: "text.magnifyingglass"
-        case .showUsage: "dollarsign.circle"
-        case .showShortcuts: "keyboard"
-        case .showNotebook: "text.book.closed"
-        case .openSettings: "gearshape"
-        }
-    }
-
-    var shortcut: String {
-        switch self {
-        case .newSession: "⌘N"
-        case .clearChat: "⌘⇧K"
-        case .clearSessions: ""
-        case .reviewChanges: "⌘R"
-        case .createCheckpoint: "⌘S"
-        case .askMode: "⌥A"
-        case .workMode: "⌥W"
-        case .planMode: "⌥P"
-        case .grillMode: "⌥G"
-        case .showShortcuts: "⌘/"
-        case .showNotebook: "⇧⌘9"
-        case .searchConversations: "⇧⌘F"
-        case .chooseWorkspace, .newWorkspace, .browseModels, .refreshModels,
-             .exportSession, .permissions, .showUsage, .openSettings: ""
-        }
-    }
-}
-
-@MainActor
-/// Coalesces work onto the display's own refresh, without trusting it to tick.
-///
-/// A `CADisplayLink` stops delivering frames whenever its display does: asleep,
-/// disconnected, or a Mac running with the lid shut and nothing attached. The
-/// link is not cancelled when that happens and reports no error — it simply
-/// goes quiet. A flush waiting on the next frame would then wait forever, and
-/// because a pending request suppresses further ones, streamed text stops
-/// appearing until something calls the flush directly.
-///
-/// So every request also arms a watchdog. Whichever arrives first wins: the
-/// frame on a live display, the watchdog on a dark one. That bounds how long a
-/// flush can be deferred without giving up display synchronisation when there
-/// is a display to synchronise with.
-///
-/// Not private, so the tests can exercise the no-frames path that a sleeping
-/// display produces and a test machine cannot otherwise reproduce.
-final class DisplaySynchronizedFlushDriver: NSObject {
-    /// How long to wait for a frame before flushing anyway. Longer than a frame
-    /// at any refresh rate a real display runs at — 40ms covers 25Hz and below,
-    /// so a live link virtually always wins — and short enough that a dark
-    /// display costs a barely perceptible delay rather than a freeze.
-    static let frameDeadlineMilliseconds = 40
-
-    private let callback: () -> Void
-    private let synchronizesWithDisplay: Bool
-    private var displayLink: CADisplayLink?
-    private var watchdog: DispatchWorkItem?
-    private var pending = false
-
-    init(synchronizesWithDisplay: Bool = true, callback: @escaping () -> Void) {
-        self.synchronizesWithDisplay = synchronizesWithDisplay
-        self.callback = callback
-    }
-
-    func request() {
-        guard !pending else { return }
-        pending = true
-        if synchronizesWithDisplay {
-            if displayLink == nil,
-               let source = NSApplication.shared.keyWindow?.screen ?? NSScreen.main
-            {
-                let link = source.displayLink(target: self, selector: #selector(displayTick(_:)))
-                link.add(to: .main, forMode: .common)
-                link.isPaused = true
-                displayLink = link
-            }
-            displayLink?.isPaused = false
-        }
-        armWatchdog()
-    }
-
-    func cancelPending() {
-        pending = false
-        displayLink?.isPaused = true
-        watchdog?.cancel()
-        watchdog = nil
-    }
-
-    func invalidate() {
-        pending = false
-        displayLink?.invalidate()
-        displayLink = nil
-        watchdog?.cancel()
-        watchdog = nil
-    }
-
-    private func armWatchdog() {
-        watchdog?.cancel()
-        let work = DispatchWorkItem { [weak self] in self?.frameNeverCame() }
-        watchdog = work
-        DispatchQueue.main.asyncAfter(
-            deadline: .now() + .milliseconds(Self.frameDeadlineMilliseconds),
-            execute: work
-        )
-    }
-
-    @objc private func displayTick(_ link: CADisplayLink) {
-        fire()
-    }
-
-    private func frameNeverCame() {
-        // Whatever silenced it — the display slept, or the screen it was built
-        // against went away — this link is no longer a clock worth waiting on.
-        // Drop it so the next request builds one against whatever display
-        // exists by then, and the app recovers on its own when one comes back.
-        displayLink?.invalidate()
-        displayLink = nil
-        fire()
-    }
-
-    private func fire() {
-        guard pending else { return }
-        pending = false
-        displayLink?.isPaused = true
-        watchdog?.cancel()
-        watchdog = nil
-        callback()
-    }
-}
-
 private struct ModelRoutingPreparedTurn {
     let id = UUID()
     let routeID: String
@@ -16876,8 +16567,4 @@ private struct ContextLoadResult: Sendable {
 private struct ChatAttachmentLoadResult: Sendable {
     let attachments: [ChatAttachment]
     let notice: String?
-}
-
-extension String {
-    var nilIfEmpty: String? { isEmpty ? nil : self }
 }
