+++ Locus/StringUtilities.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+extension String {
+    var nilIfEmpty: String? { isEmpty ? nil : self }
+}
